python-service/requirements.txt
python-docx>=1.1.0
fastapi>=0.115.0
uvicorn>=0.34.0
